preprocessing/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
pandas==2.0.3
networkx==3.1
fa2==0.3.5
cdlib==0.3.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
scipy==1.10.1
torch==2.1.2+cpu
torch_geometric==2.4.0
torchaudio==2.1.2+cpu
torchvision==0.16.2+cpu
transformers==4.39.0
sentence-transformers==2.2.2
umap-learn==0.5.3
Werkzeug==2.3.6
emoji==2.10.1
fasttext-langdetect==1.0.5
nltk==3.8.1
pysolr==3.9.0
numpy
python-dotenv==1.0.1
requests==2.31.0
huggingface-hub==0.21.4
protobuf==5.26.0
gunicorn==21.2.0
tensorflow==2.14.0
tensorflow-probability==0.22.1